Create provision-azure-infrastructure.md
I may have to tweak my suggestions a bit more, but wanted to commit these to advance the process. This is a new process for me too, so I'm not sure if the level of changes I made is appropriate for this new review process. 

I moved up the item that said to set the number of replicas to 1, to ensure that the YAML file is modified before being applied to the cluster. The steps assume that people have the YAML file but we don't mention the delegate installer -- do people know how they should have gotten the delegate YAML file?

Looking forward to seeing this on the site. Thank you.