Update install-delegate.md
Fixed one broken link. For an editing pass, I would remove the link that was broken since it is applied to the first two words in an article about installing the delegate and the reader has no way of knowing why they're being linked away since there is no "For more information". The link can come at the end of the intro and be prefaced with a "For an overview of delegates," or something like that. The link was changed to: https://developer.harness.io/docs/platform/Delegates/get-started-with-delegates/delegates-overview