Update nodejs-firstscan.md
I updated a paragraph in the doc. The original sentence read like this:
Since Harness STO can orchestrate multiple security scanning tools, setup requires a few details such as what type of scanner will be used and location information for what will be scanned. These are entered as key value pairs. Can enter the following key value pairs:

I changed it to:
Since Harness STO can orchestrate multiple security scanning tools, setup of it requires a few details, such as what type of scanner is planned to be utilized and the location of information for what will be scanned. These are entered as key-value pairs as follows: 

This got rid of a run-on sentence as well as removing a redundant statement in the original paragraph.