Update self-managed-enterprise-edition.md
Draft of NG SMP notes for 2/6/23 or thereabouts. There are some identifiers for which I could not find a written summary thus far. Those are justified at the margin so should be clear. I did a bit of post-migration cleanup in this file but not enough time to go through the entire file.