Update harness-git-ops-application-set-tutorial.md (#648)
* Update harness-git-ops-application-set-tutorial.md

I was asked to add an explanation of wave deployments to the PR Pipeline doc by Cat Palmer to support a blog article. Cat wants to ensure that someone searching for "wave deployments" in our docs finds it discussed as part of GitOps. It belongs in the PR Pipeline topic as that the blog article discusses wave deployments in the GitOps context.

blog draft (not sure who wrote it): https://docs.google.com/document/d/1gUoabLk8bN3odoAN-dh_-6Wejcp6PNaeutTPvPPOgOY/edit

* Apply suggestions from code review

Co-authored-by: April M <36110273+aimurphy@users.noreply.github.com>

---------

Co-authored-by: April M <36110273+aimurphy@users.noreply.github.com>