Update define-a-docker-build-infrastructure.md
The current document for installing docker runner does not cover an error on MacOS that will block users. Although it's possible for users to search on the internet and find the same solution, but I believe many users will just give up. 
Here add the solution to our document directly.