Update cloud-formation-how-tos.md (#1587)
if the URL of the current page has a trailing / all the links on the page that use the format [title](./page-name) will fail. 

This is because of how our web server handles trailing /.

Ravi asked me to change the links on this page. The problem with our web server still exisits.