Add details of Barriers usage with Matrix. (#3718)
* Add details of Barriers usage with Matrix.

* Edits to the content

LGTM. Added a reference to the important notes section. Will need to test the reference in the PR environment. 

(Updates did not show up on my local system even after switching to this branch - git pull says up to date, and I can see the updates with a git diff at the bash prompt, but the local files are not updated, so I'll have to test in the PR env.)

* Additional details based on walkthrough

* Apply suggestions from code review

doc review

---------

Co-authored-by: Deepak Bolangady <deepak.bolangady@harness.io>
Co-authored-by: Michael Cretzman <58786311+michaelcretzman@users.noreply.github.com>