[PL-46758] Add info in docs regarding the SYSTEM updates
Add info in docs regarding the SYSTEM updates done on delegated displayed on audit logs.

"User information is not propagated when a Delegate agent is started from external sources. As a result, all subsequent operations by the Delegate agent will be recorded under the "SYSTEM" user. This includes the creation, deletion, and upsertion of delegates. This behavior is expected during the platform's upgrade process."