Update git-hub-connector-settings-reference.md
A common error mode for users is giving the wrong key, and the error message is not always clear that the format is the problem. This should be front and center in all concerned docs as this has been an issue in FG and NextGen that customers hit and don't realize the problem. 

Please revise as necessary for clarity or consistency in format