add Javascript SDK (under Javascript and Typescript language categories)
This commit is to preserve work until now. I will next reorganize the way React components are imported in the AdvancedFeatureFlagEvaluations and AdvancedUseCases index files. This is to allow correct rendering of the right-hand side menu. The side menu is rendered before the props are evaluated so the current approach has a timing issue.