Update fmeData.ts removed Gemini Code Assist
Per announcement in #ask-split-fme-product 3/4/25: 
Gemini Code Assist development is blocked and will be removed from the roadmap. It requires Harness to conform a specific google auth requirement that the platform does not currently support – and is not on the roadmap. When/if this changes, the roadmap will be updated accordingly.