Update single-sign-on-saml.md clarify there is no download function
The actions menu let's you "view IdP metadata" which brings it up in a browser window.  The user must then copy that and paste it by hand into a new file and save it.  Using the word "download" here runs the risk of leading them to the other menu option, which is "Download certificate" and not what is needed here.