Update sso-for-admins.md clarify SSO done in Harness and IdP
A customer interpreted the existing directions to imply they did all the work in just their IdP and then they should notify us to complete the work on the Harness side.  

Making a simple wording change to make it clear that they must log into both Harness and their IdP to complete the directions below.